+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ### Unreleased
+
+### 2.1.15
+- Remove no-used initState in header reducer
+- Update twreporter-redux
+- Update topics: compose data in mapStateToProps instead
+- Update topics: fix fetch data requests
+
 ### 2.1.14
 - Add topics from twreporter-react-listing-components